requirements.txt
elevenlabs==1.13.1
langchain==0.3.9
langchain_google_genai==2.0.6
numpy==2.1.3
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.68
opencv_python_headless==4.8.0.74
pandas==1.5.3
python-dotenv==1.0.1
scikit_learn==1.5.2
supervision==0.25.0
ultralytics==8.2.2
